feat(ws): move files, implement broadcast
