+++ apps/backend/plugins/websocket/adapters/getConnections.ts
@@ -1,0 +1,8 @@
+import { Connections } from '../../../types';
+
+const connections: Connections = new Map();
+
+export function getConnections(): Connections {
+	return connections;
+}
+
