+++ apps/backend/plugins/websocket/adapters/getRooms.ts
@@ -1,0 +1,7 @@
+import { Rooms } from "../../../types";
+
+const rooms: Rooms = new Map();
+
+export function getRooms() {
+	return rooms;
+}
